refactor: renaming components
