--- src/components/Headers/Title.jsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components'
-
-export const Title = styled.h1`
-  display: flex;
-  justify-content: flex-start;
-`
