--- src/components/Headers/index.js
@@ -1,3 +1,0 @@
-export { default as Header } from './Header'
-export * from './Header'
-export * from './Title'
