+++ src/components/index.js
@@ -1,10 +1,10 @@
 export { default as Bio } from './Bio'
 export * from './BlogExcerpt'
-export * from './Headers'
+export * from './Header'
 export * from './Search'
 export { default as Layout } from './Layout'
 export * from './PostDetails'
-export * from './PostTitleLink'
+export * from './PostTitle'
 export * from './Navigation'
 export { default as SEO } from './SEO'
 export * from './StyledLink'
